+++ workspaces/docs/src/pages/docs-form/index.ts
@@ -1,0 +1,9 @@
+import { useEffect } from 'react';
+
+export default function DocsForm() {
+  useEffect(() => {
+    window.location.href = '/usetheform'; // Redirect to the home page
+  }, []);
+
+  return null
+}