+++ workspaces/usetheform/src/utils/formUtils.js
@@ -1,14 +1,18 @@
 import { STATUS, FORM_VALIDATION_LABEL } from "./constants";
 
-export const createForm = (state = {}) => ({
-  state,
-  isValid: true,
-  status: STATUS.READY,
-  pristine: true,
-  isSubmitting: false,
-  submitAttempts: 0,
-  submitted: 0
-});
+export const createForm = (state = {}, formStore) => {
+  const formState = {
+    state,
+    isValid: true,
+    status: STATUS.READY,
+    pristine: true,
+    isSubmitting: false,
+    submitAttempts: 0,
+    submitted: 0
+  };
+  formStore.update(formState, false);
+  return formState;
+};
 
 export const getValueByPath = (path, obj, separator = "/") => {
   if (path === FORM_VALIDATION_LABEL) {